Update main.py
countries (2).csv is the same as the data but with a header added to it so that the Andorra data line is not considered a header and thus not removed from the lists of data for latitude, longitude, country, native spelling, and native language. Note-The file is a copy of the original and not the original itself.